Move `kubetools_client` -> `kubetools`.
