+++ kubetools/cli/__main__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+
+from kubetools.cli import cli_bootstrap
+from kubetools.main import run_cli
+# Import click command groups
+from kubetools.cli import deploy, show  # noqa: F401, I100
+
+
+run_cli(cli_bootstrap)
