+++ setup.cfg
@@ -3,4 +3,4 @@ inline-quotes = single
 multiline-quotes = single
 docstring-quotes = single
 import-order-style = edited
-application-import-names = kubetools_client
+application-import-names = kubetools
